refactor: ほしいものリストのコンポーネント名を変更
